// src/dto/User.dto.ts
export interface IGeo {
  lat: string
  lng: string
}

export interface IAdress {
  street: string
  suite: string
  city: string
  zipcode: string
  geo: IGeo
}

export interface ICompany {
  name: string
  catchPhrase: string
  bs: string
}

export interface IUser {
  id: number
  name: string
  username: string
  email: string
  address: IAdress
  zipcode: string
  phone: number | string
  website: string
  company: ICompany
}